Clean up core interfaces, hide some apis, move some others around
