+++ webauthn-authenticator-rs/src/softpasskey.rs
@@ -549,7 +549,7 @@ mod tests {
         let name = "william";
 
         let (chal, reg_state) = wan
-            .generate_challenge_register_options(
+            .generate_challenge_register(
                 &unique_id,
                 name,
                 name,
@@ -578,7 +578,7 @@ mod tests {
         let cred = wan.register_credential(&r, &reg_state, None).unwrap();
 
         let (chal, auth_state) = wan
-            .generate_challenge_authenticate(vec![cred], None)
+            .generate_challenge_authenticate(vec![cred], None, None, None)
             .unwrap();
 
         let r = wa
