+++ webauthn-rs-core/src/interface.rs
@@ -38,7 +38,7 @@ pub struct RegistrationState {
     pub(crate) require_resident_key: bool,
     pub(crate) authenticator_attachment: Option<AuthenticatorAttachment>,
     pub(crate) extensions: RequestRegistrationExtensions,
-    pub(crate) experimental_allow_passkeys: bool,
+    pub(crate) experimental_allow_synchronised_authenticators: bool,
 }
 
 /// The in progress state of an authentication attempt. You must persist this associated to the UserID
